fix issues, move some files
