+++ src/commands/auth/login.ts
@@ -18,11 +18,11 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
+import { recordConnectionFromAuth } from '@/core/auth/auth-connection-recorder.ts';
 import { isSonarQubeCloud } from '@/core/auth/auth-resolver.ts';
 import { type BrowserAuthResult, generateTokenViaBrowser } from '@/core/auth/token.ts';
 import { CommandFailedError, InvalidOptionError } from '@/core/command-error.ts';
 import { SONARCLOUD_URL, SONARCLOUD_US_URL } from '@/core/config-constants.ts';
-import { recordConnectionFromAuth } from '@/core/host/auth-connection-recorder.ts';
 import {
   deleteStaleTokens,
   getToken as getKeystoreToken,
