+++ src/commands/integrate/_common/context-augmentation-dependency.ts
@@ -33,8 +33,8 @@ import {
   installContextAugmentationBinary,
   resolveContextAugmentationBinaryPath,
 } from '@/core/host/install/context-augmentation.ts';
+import { CONTEXT_AUGMENTATION_BINARY_NAME } from '@/core/host/install/install-types.ts';
 import { SONAR_CONTEXT_AUGMENTATION_VERSION } from '@/core/host/install/signatures.ts';
-import { CONTEXT_AUGMENTATION_BINARY_NAME } from '@/core/host/install-types.ts';
 
 import { stopAllContextAugmentationTools } from './context-augmentation.ts';
