+++ src/core/framework/dependencies/sonarsource-binary.ts
@@ -25,6 +25,7 @@ import {
   removeBinary,
   resolveBinaryPath,
 } from '@/core/host/install/binary.ts';
+import { SCA_SCANNER_BINARY_NAME, SECRETS_BINARY_NAME } from '@/core/host/install/install-types.ts';
 import {
   SCA_SCANNER_CLI_SIGNATURES,
   SCA_SCANNER_CLI_VERSION,
@@ -32,7 +33,6 @@ import {
   SONAR_SECRETS_VERSION,
   SONARSOURCE_PUBLIC_KEY,
 } from '@/core/host/install/signatures.ts';
-import { SCA_SCANNER_BINARY_NAME, SECRETS_BINARY_NAME } from '@/core/host/install-types.ts';
 
 import type {
   DependencyInstallContext,
