+++ src/core/host/install/install-types.ts
