+++ src/core/host/install/secrets.ts
@@ -21,12 +21,12 @@
 // sonar-secrets install: thin wrapper over the generic binary install pipeline.
 
 import { SONAR_SECRETS_DIST_PREFIX } from '@/core/config-constants.ts';
+import { type PlatformInfo, SECRETS_BINARY_NAME } from '@/core/host/install/install-types.ts';
 import {
   SONAR_SECRETS_SIGNATURES,
   SONAR_SECRETS_VERSION,
   SONARSOURCE_PUBLIC_KEY,
 } from '@/core/host/install/signatures.ts';
-import { type PlatformInfo, SECRETS_BINARY_NAME } from '@/core/host/install-types.ts';
 import { discreetSuccess, type OutputChannel } from '@/core/ui';
 
 import {
