+++ src/core/telemetry/identity-fetch.ts
