+++ tests/e2e/context/cag-edge-cases.test.ts
@@ -32,8 +32,8 @@ import { afterEach, beforeEach, describe, expect, it, setDefaultTimeout } from '
 import { CLAUDE_INTEGRATION_ID } from '@/commands/integrate/claude/declaration.ts';
 import { detectPlatform } from '@/core/host/environment/platform-detector.ts';
 import { buildLocalCagBinaryName } from '@/core/host/install/context-augmentation.ts';
+import { CONTEXT_AUGMENTATION_BINARY_NAME } from '@/core/host/install/install-types.ts';
 import { SONAR_CONTEXT_AUGMENTATION_VERSION } from '@/core/host/install/signatures.ts';
-import { CONTEXT_AUGMENTATION_BINARY_NAME } from '@/core/host/install-types.ts';
 import type { CliState } from '@/core/state/state.ts';
 
 import { version as CURRENT_CLI_VERSION } from '../../../package.json';
