+++ tests/unit/commands/update/sonarsource-releases.test.ts
@@ -21,7 +21,7 @@
 import { describe, expect, it } from 'bun:test';
 
 import { SONAR_SECRETS_DIST_PREFIX, SONARSOURCE_BINARIES_URL } from '@/core/config-constants.ts';
-import { buildDownloadUrl } from '@/core/host/sonarsource-releases.ts';
+import { buildDownloadUrl } from '@/core/host/install/sonarsource-releases.ts';
 
 const SECRETS = 'sonar-secrets';
