+++ tests/unit/core/auth/auth-resolver.test.ts
@@ -22,6 +22,7 @@
 
 import { afterEach, beforeEach, describe, expect, it, spyOn } from 'bun:test';
 
+import * as authConnectionRecorder from '@/core/auth/auth-connection-recorder.ts';
 import {
   cloudRegionFromUrl,
   ENV_SERVER,
@@ -31,7 +32,6 @@ import {
   resolveAuth,
   resolveFromEndpoint,
 } from '@/core/auth/auth-resolver.ts';
-import * as authConnectionRecorder from '@/core/host/auth-connection-recorder.ts';
 import { getDefaultState } from '@/core/state/state.ts';
 import * as stateRepository from '@/core/state/state-repository.ts';
 import { clearMockUiCalls, getMockUiCalls, setMockUi } from '@/core/ui';
