+++ tests/unit/core/host/distribution.test.ts
@@ -20,10 +20,7 @@
 
 import { describe, expect, it } from 'bun:test';
 
-import {
-  resolveDistribution,
-  resolveDistributionConfig,
-} from '@/core/host/environment/distribution.ts';
+import { resolveDistribution, resolveDistributionConfig } from '@/core/host/distribution.ts';
 
 describe('distribution', () => {
   it('defaults to standalone when no distribution is provided', () => {
