+++ tests/unit/core/telemetry/identity.test.ts
@@ -35,7 +35,6 @@ import { afterEach, beforeEach, describe, expect, it, spyOn } from 'bun:test';
 import type { ResolvedAuth } from '@/core/auth/auth-resolver.ts';
 import * as authResolver from '@/core/auth/auth-resolver.ts';
 import { ENV_SONAR_USER_HOME, getTelemetryDir } from '@/core/config-constants.ts';
-import { resolveTelemetryIdentity } from '@/core/host/identity-fetch.ts';
 import type { AuthConnection } from '@/core/state/state.ts';
 import {
   identityFromConnection,
@@ -43,6 +42,7 @@ import {
   resolveCommandTelemetryIdentity,
   resolveStoreEventTelemetryIdentitySafely,
 } from '@/core/telemetry/identity.ts';
+import { resolveTelemetryIdentity } from '@/core/telemetry/identity-fetch.ts';
 
 import { mockIdentityGetSafe } from './identity-api-mock.ts';
