+++ tests/unit/core/telemetry/telemetry-events.test.ts
@@ -35,8 +35,8 @@ import { scanAndEmitSecrets } from '@/commands/analyze/secrets.ts';
 import type { ResolvedAuth } from '@/core/auth/auth-resolver.ts';
 import { ENV_SONAR_USER_HOME, TELEMETRY_ENDPOINT } from '@/core/config-constants.ts';
 import * as networkConfig from '@/core/host/connectivity/network-config.ts';
+import { DISTRIBUTION } from '@/core/host/distribution.ts';
 import * as agentDetector from '@/core/host/environment/agent-detector.ts';
-import { DISTRIBUTION } from '@/core/host/environment/distribution.ts';
 import type { SpawnResult } from '@/core/process/process.ts';
 import * as fetchGuardedModule from '@/core/server/fetch-guarded.ts';
 import type {
