+++ tests/unit/core/update/binary-refresh.test.ts
@@ -20,9 +20,9 @@
 
 import { afterEach, beforeEach, describe, expect, it, Mock, spyOn } from 'bun:test';
 
+import { SCA_SCANNER_BINARY_NAME } from '@/core/host/install/install-types.ts';
 import * as scaScannerInstall from '@/core/host/install/sca-scanner.ts';
 import * as secretsInstall from '@/core/host/install/secrets.ts';
-import { SCA_SCANNER_BINARY_NAME } from '@/core/host/install-types.ts';
 import type { CliState } from '@/core/state/state.ts';
 import { getDefaultState } from '@/core/state/state.ts';
 import * as stateRepository from '@/core/state/state-repository.ts';
